to-do-get-kaggle-dataset/requirements.txt
kagglehub[pandas-datasets]
pymongo[srv]
python-dotenv